+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - Also remove standard fields when running **hardis:project:clean:standarditems**
+- New Deployment tips
+  - Wrong api Version of a Metadata
+  - Unknown user
 - Upgrade to MegaLinter v7
 
 ## [3.16.0] 2022-05-24